add default preferences handling
