+++ sqlboiler.toml
@@ -9,4 +9,4 @@ host = "localhost"
 port = 26257
 user = "root"
 sslmode = "disable"
-blacklist = ["goose_db_version"]
+blacklist = ["goose_db_version", "notification_defaults"]
